scripts/bash_aliases: man_lsfunc(): Fix some corner cases
Although 'int foo(...);' is not a valid C prototype, we use it in
one manual page: sigreturn(2).  Fix man_lsfunc() to find that
prototype.

When a syscall(SYS_...) function doesn't have any arguments
(except for the SYS_* constant), there's not a comma between the
parentheses.  Adapt the regex for that.

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>