ctime.3: SYNOPSIS: Use VLA notation for [as]ctime_r() buffer
As N2417 (a proposal for C2x that was not accepted) suggests.
This syntax is very informative, and also, if used by library
implementers, can improve static analysis.

Since it is backwards compatible with pointer syntax, we can do
this.  Also, the reason for not being accepted, as Jens noted,
was incompatibility with C++.  But IMO, that's not something that
we should care too much.  Users of other languages should know
what to expect when reading C code, and are expected to know C/C++
incompatibilities.

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>
Cc: Jens Gustedt <jens.gustedt@loria.fr>
Cc: Glibc <libc-alpha@sourceware.org>