pow.3: RETURN VALUE: Reorder special cases as in the standard
Use the following order:

- Success
- Over/Underflow
- Special cases specified by the standard (in the same order)
- Other special cases not specified by the standard.

This way it's easier to compare this document against the standard
to check for any mistakes or differences.

I also added a comment at the beginning of each case, to more
clearly delimit special cases, with a syntax that is easy to read
(a mix of C, mathematical language, ...).  We may want to use
those as tagged paragraphs.

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>
Cc: M. Welinder <mwelinder@gmail.com>