usleep.3: Remove incorrect portability note
That note against using useconds_t if portability is important was
probably true before POSIX, but since POSIX specifies the type in
<sys/types.h>, as recently documented in useconds_t(3), the type
should be fine to use portably.  Moreover, using it helps document
the code, so we better completely remove that warning.

Add a SEE ALSO entry for useconds_t(3).

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>