bash_aliases: grep_*(): Optimize
Using grep first to reduce the files on which pcregrep is run
optimizes considerably the performance of these functions.

If a file doesn't contain the name of a function,
which can easily be checked with grep,
it won't possibly contain the function prototype or definition,
which is much slower to search,
since it implies multiline pcregrep search.

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>