access.2: Document EPERM error code
[Todd]:

If a file/directory has the immutable flag set, access/faccessat
will return EPERM if write access is requested.  This is not
currently documented, presumably because it is Linux-specific
(though BSD has similar behavior that is documented).  Perhaps
something like the following in the ERRORS section.

EPERM	Write permission was requested and pathname has the
	immutable flag set.  (See also ioctl_iflags(2).)

Also perhaps add ioctl_iflags(2) to the SEE ALSO section.

This came up in reference to:
<https://github.com/sudo-project/sudo/issues/122>

Reported-by: "Todd C. Miller" <Todd.Miller@sudo.ws>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>