capabilities.7: improve internal references
Trying to make this man page easier to navigate.

Fix a few cases of "see above/below" without a specific reference to a
subsection by quoting the subsection name (making it easier to look it up).
Use the same formatting rule as used by some of the other existing
references, i.e. italicise it.

For uniformity, remove words such as "the subsection" and "under", using
"(see|described in) <subsection title> (above|below)" template.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>