Makefile, etc/groff/tmac/deadly.tmac: Add lint and lint-groff targets
'lint' is a metatarget that triggers all lint-* targets (currently
that is just 'lint-groff').
'lint-groff' is a target that runs groff -rCHECKSTYLE to check
manual pages' groff(7) correctness.

etc/groff/tmac/deadly.tmac is a file written by Branden, to make
groff(1) abort when it finds a style problem, which then causes
make(1) to also abort.

CC: "G. Branden Robinson" <g.branden.robinson@gmail.com>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>