uri.7: The term URL is deprecated, in favor of URI
See RFCs 4395 (obsoleted by 7595) and 7595.

So, since URI and URL have been used as synonims and it wasn't
clear to most people what the difference was, the RFCs seem to
have explicitly avoided URL, and now use URI as the only term,
which still means what it meant (so now URL is just a synonim for
URI).

This commit replaces (almost all) occurences of URL by URI,
except when it is referring to the old term itself.  Keep some
legacy info just for readers to understand this.

Cc: NGINX Unit <unit@nginx.org>
Cc: Jakub Wilk <jwilk@jwilk.net>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>