mmap.2: wfix
Change "supported only since" to "support was added in".

Previously, this said, "MAP_POPULATE is supported for private mappings only
since Linux 2.6.23."  This could be parsed as "MAP_POPULATE is supported
for private mappings only [not shared mappings] since Linux 2.6.23."
Someone (a native English speaker) tried to tell me this.

Change to "Support for MAP_POPULATE in conjunction with private mappings
was added in Linux 2.6.23."  This is unambiguous, and "support was
added" is more common in other man pages.

There are more instances of "supported only since", but they seem
clearer.  I could send a separate patch for them if desired.

Signed-off-by: Jesse Rosenstock <jmr@google.com>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>