Various pages: wfix
Replace "shall" wordings by simpler text, to not resemble a
standards document.

When saying something like "According to POSIX/ISO C, ...", if
the mentioned standard is the same one described in STANDARDS,
remove that text (it is redundant).  However, if ISO C is the
main standard that describes a type, but POSIX applies further
restrictions, keep that text.

Remove pedantic notes that may only be interesting to compiler
writers.

Reported-by: Ingo Schwarze <schwarze@openbsd.org>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>