append_COLOPHON.sh, remove_COLOPHON.sh: Remove COLOPHON section.
Users should use their distribution's standard tools to know where
a given page comes from and to know how to report a bug.  That'll
create a small barrier to bug reporters, but that'll hopefully
help improve the quality of the reports.

It'll also make the pages cleaner, by removing stuff that is
obvious to those who know how to use their tools.

We still keep part of the information of the COLOPHON,
specifically the project name and the version, in the .TH line.

Reported-by: Ingo Schwarze <schwarze@openbsd.org>
Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>