Revert "lint-man.mk: Use ascii instead of utf8"
This reverts commit f1a71d4a836d48f81832415626d27521854f85c6.

The issues motivating that patch were not real, and I guess I had
another problem, probably caused by me (EBCAK).  Since I can't
reproduce the issues that lead to the change, and it's preferrable
to test exactly what users will be reading on their terminals,
let's use UTF-8 again for the tests, which is what most users will
be reading.

Reported-by: Ralph Corderoy <ralph@inputplus.co.uk>