Various pages: EXAMPLES: Use <err.h> functions
When reporting errors, prefer using conventional <err.h>
functions, rather than home-made macros.  Home-made macros are
necessary for portable programs, so in documentation about
portable functions, keep using the macros.  However, in the
documentation for functions that are only available in GNU/Linux
and/or BSD systems, prefer <err.h>, which is available in both
systems.

Do the same in example programs documenting portable functions but
in which we already require _GNU_SOURCE for some other reason.

Signed-off-by: Alex Colomar <alx.manpages@gmail.com>