Various pages: EXAMPLES: Use unsigned types for loop iterators
Looping with unsigned types is safer.  See the link below.

When the iterators are used for accessing an array, use size_t;
otherwise, use the most appropriate unsigned type, which in most
cases is just 'unsigned int'.

Also adjust other variables that have to interact with the
iterators, to avoid comparison of integers of different
signedness.

Link: <https://gustedt.wordpress.com/2013/07/15/a-praise-of-size_t-and-other-unsigned-types/>
Cc: Jens Gustedt <jens.gustedt@inria.fr>
Signed-off-by: Alex Colomar <alx.manpages@gmail.com>