dist.mk, All pages: .TH: Generate date at 'make dist'
Replace the date by a placeholder (date) in the repo, as we're
doing with the version (unreleased).  It will be filled when the
tarball is generated with 'make dist' (or equivalent) with the
date of the most recent git commit that modifies the page (as was
done previously by update_timestamps.sh, which has been removed).

Scripted change (mostly):

$ find man* -type f \
| xargs sed -Ei '/^\.TH /s/.TH +([^ ]+ +[^ ]+) +[^ ]+ +(.*)/.TH \1 (date) \2/'

Signed-off-by: Alejandro Colomar <alx@kernel.org>