README: Rewrite and expand
Let's start making the repository self-contained.  The README
should be enough to guide a user into everything that has to do
with the project.  The website should be a secondary source, for
those who don't know about the repository yet.

I merged some of the information found in the .Announce file,
which I'll remove.

Signed-off-by: Alejandro Colomar <alx@kernel.org>