Makefile: Remove installdirs* and builddirs* targets
Since our makefiles correctly state the dependencies of all
targets, and allow fully-parallel builds and installs, there's no
reason at all to allow creating the directories separately.  That
was an old recommendation for when makefiles were buggy, and you
couldn't fully trust in them not having races or incorrect
dependencies.

Signed-off-by: Alejandro Colomar <alx@kernel.org>