dist.mk: Force adding version to tarball man pagesat 'make dist'
When creating a tarball with 'make dist', force the addition of
the date and version to the pages.  Otherwise, if a tarball had
already been created from a previous commit, and the page hasn't
been touched, make will consider the one in the tarball
up-to-date, and will keep the old version.

Signed-off-by: Alejandro Colomar <alx@kernel.org>