keyrings.7: tfix hexadecimal number prefix
In the DESCRIPTION, where the Permissions column of /proc/keys is
explained, the 0x02 value was written as Ox02. In case one cannot spot
the difference: the first character was letter o (U+004F), while it
should be digit zero (U+0030).

It seems the source tree does not contain any other "Ox" instances
except at the beginning of "Oxford".

Signed-off-by: Grzegorz Szymaszek <gszymaszek@short.pl>
Signed-off-by: Alejandro Colomar <alx@kernel.org>