string.3, strncat.3: Undeprecate strncat(3)
After some investigation, I found a case where this function is useful:
Concatenating an unterminated string into a string.  It's not an ideal
API for that, but there's no other API that does it.

The closest thing, and something that some projects use instead of
strncat(3), is calling mempcpy(3) directly.  However mempcpy(3) isn't
ideal either (it's faster; just that).  It even requires a multiline
pattern to use correctly, which is a source of bugs.

So, suggest using a custom alternative that needs to be defined by the
programmer, which handles all the subtle details much better than any
of the conventional functions: ustr2stpe().

Signed-off-by: Alejandro Colomar <alx@kernel.org>