strncat.3: Overrunning the dest buffer is not that bad
With the right function call, that is, one that always copies the same
amount of bytes, and is so simple that can be inlined, the behavior
will be consistent enough to be warned by the compiler in most cases of
overrun, and crash quite consistently in the remaining.

Prefer simplicity over correctness, so suggest the simpler ustr2stp().

Signed-off-by: Alejandro Colomar <alx@kernel.org>