suffixes.7: Improve table format
Apparently the original author(s) of this table did not know how to
start a table entry in the first column of a tbl(1) table with a dot.
(If you try, the *roff formatter will interpret it as a control line,
and try to invoke a request or call a macro.)

Start every row of the table with the *roff dummy character `\&` instead
of leading space.  Not only is this more idiomatic, but it recovers some
of the line length for content.

This patch does not attempt to correct any errors in the table contents,
nor bring it up to date from its year 2000 vintage.

Signed-off-by: Alejandro Colomar <alx@kernel.org>