scanf.3, sscanf.3, vsscanf.3: Split the page, one for strings and one for FILEs
From these functions, the ones that read from a FILE* are very difficult
--if not impossible-- to use correctly.  Let's split the page into two,
so that we give the impression that sscanf(3) is the first class
citizen, and the others are just historic artifacts that are kept for
backwards-compatibility reasons.

FILE* variants are now in a page which clearly tells the reader to look
for other ways to read input.

Link: <https://lore.kernel.org/linux-man/633629bd-753c-3097-9896-2491a0b0f1a2@gmail.com/T/>
Cc: Ian Abbott <abbotti@mev.co.uk>
Cc: Zack Weinberg <zack@owlfolio.org>
Signed-off-by: Alejandro Colomar <alx@kernel.org>