localedef.1: ffix
Stop manipulating adjustment and hyphenation.

Forcibly re-enabling adjustment to both margins after the synopsis does
not respect user configuration of adjustment.  There _is_ a portable way
to save the adjustment mode, via the .j register and a copy of it, but
doing so requires even more usage of low-level requests that are
discouraged in man page writing.

The latter is incorrect for use with groff(1) since '.hy' does not
restore the previous hyphenation mode but sets it to 1, which is not
appropriate for the English-language hyphenation patterns groff uses.
(Also, AT&T man(7) used a hyphenation mode of 14.)

Neither of these requests is respectful of user configuration of
adjustment or hyphenation enablement.  Features in the forthcoming
groff 1.23 will make these easier for users to manipulate to their
preference.

(mandoc(1) does not support configurable adjustment or hyphenation, so
all of these requests are no-ops for it.)

Signed-off-by: G. Branden Robinson <g.branden.robinson@gmail.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>