CONTRIBUTING: Fix typo, there is one active maintainer
On commit "CONTRIBUTING, README, lsm: Remove mtk as maintainer"
(06e72cb1) we changed to mail only one maintainer, but the doc still
says "both maintainers".

When submitting a patch, I was confused by that fact and thought Michael
address was missing.  But after checking, it seems we just need to send
it to Alejandro, so clarify the text to match that.

Signed-off-by: Rodrigo Campos <rodrigo@sdfg.com.ar>
Signed-off-by: Alejandro Colomar <alx@kernel.org>