INSTALL, cmd.mk, install-man.mk: Support installing compressed pages
Distributions usually install compressed (.gz) pages to reduce space.
Let's support this in our build system, as a command-line variable "Z",
which is empty by default, but can be set to a file extension to append
to the page names (and the appropriate compression program will be
used).  For now, the only compression supported is ".gz".

Example:

    $ make install Z=.gz

This can be combined with LINK_PAGES, to produce compressed pages and
use symbolic links for the link pages:

    $ make install Z=.gz LINK_PAGES=symlink

Signed-off-by: Alejandro Colomar <alx@kernel.org>