*.mk: Remove unnecessary '.' after directory names (but keep the '/')
I used it for some reason I don't remember, probably because I did
something wrong, and didn't know how to do it right.  I've tried now
without it, and it's working, so let's just remove it.

While we don't want trailing slashes in directory variables, we want
them in targets, so we can distinguish directory targets.

Signed-off-by: Alejandro Colomar <alx@kernel.org>